pyproject.toml
[tool.poetry]
name = "client_broker"
version = "0.1.0"
description = ""
authors = ["Andrés <andres.sandoval@alaya.cl>"]

[tool.poetry.dependencies]
python = "^3.7"
sqlalchemy = "^1.3.16"
petri = "^0.24.1"
paho-mqtt = "^1.5.0"
jinjasql = "^0.1.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.4.4"
ipython = "^7.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
